Make Vim's tmp file settings safer
The use of '~' in constant strings to refer to the user's home directory
is replaced with "$HOME" and execute commands.  This is more robust as
"$HOME" can be always be used to refer to the home directory whereas '~'
only refer to the directory in certain cases (see `:h $HOME`).

Each `set` of a tmp file variable is accompanied by a call to `mkdir()`.
Those function calls create the preceding directory (along with parent
directories, as necessary) with permissions that restrict permissions to
the owner.

`undodir` and `undofile` only exist if Vim was compiled with the
`+persistent_undo` feature (see `:h 'undofile').  As such, the setting
of those variables should be conditional on the compilation setting.